feat(lms): major MVC restructure, auth revamp, and payment flow overhaul
* Replace Livewire auth with controller-based auth + Blade views
* Add full Admin/Instructor/Student module controllers and pages
* Introduce payment UI per gateway (PayPal/Stripe/Razorpay/etc.)
* Bridge API → web session in Api\PaymentController; fix guard usage
* Refactor helpers (ApiHelpers/CommonHelpers); stop double json\_decode on casted arrays
* Add services: WatermarkService, ZoomMeetingService
* Add middleware: AdminMiddleware, InstructorMiddleware, BlogVisibility, etc.
* Update models (Course/User/WatchDuration) and course casting/accessors
* Split routes into admin, instructor, student, payment, player, chat, guest, system
* Add chats table migration; update courses migration and seeders
* Large Blade addition: course player, certificate builder, blog, bootcamp, settings, layouts
* Update .gitignore, composer deps, app bootstrap/config
* Remove obsolete Livewire components and legacy layouts

BREAKING CHANGE: Auth pages/components migrated from Livewire to controllers; routes reorganized. Run `composer install`, `php artisan migrate --seed`, and clear caches.

+++ backend/course/app/Models/BootcampPurchase.php
@@ -25,10 +25,10 @@ class BootcampPurchase extends Model
     ];
 
     protected $casts = [
-        'price' => 'decimal',
-        'tax' => 'decimal',
-        'admin_revenue' => 'decimal',
-        'instructor_revenue' => 'decimal',
+        'price' => 'float',
+        'tax' => 'float',
+        'admin_revenue' => 'float',
+        'instructor_revenue' => 'float',
         'status' => 'integer',
     ];
